Isolate getUserDetails controller tests
